Removed tpfa_coupling and upwind_coupling.
The classes are in tpfa.py and upwind.py, respectively.
