+++ test/unit/test_upwind_elimination.py
@@ -10,7 +10,7 @@ from porepy.params.data import Parameters
 
 from porepy.numerics.fv import tpfa, fvutils
 from porepy.numerics.fv.source import IntegralMultiDim
-from porepy.numerics.fv.transport import upwind, upwind_coupling
+from porepy.numerics.fv.transport import upwind
 from porepy.numerics.mixed_dim import coupler, condensation
 #------------------------------------------------------------------------------#
 
@@ -141,8 +141,7 @@ class BasicsTest(unittest.TestCase):
 
         #------Transport------#
         advection_discr = upwind.Upwind(physics="transport")
-        advection_coupling_conditions = upwind_coupling.UpwindCoupling(
-            advection_discr)
+        advection_coupling_conditions = upwind.UpwindCoupling(advection_discr)
         advection_coupler = coupler.Coupler(
             advection_discr, advection_coupling_conditions)
         #U, rhs_u = advection_coupler.matrix_rhs(gb)
